ViewTutorProfile - Implement logic for displaying tutor profile.
The problem is that Parse does not let us modify the data of a user who is not logged in. This is why I had to create a new table called Ratings, that will store the username and the rating  + rating count. I will modify the Edit Profile code for adherence as well.