Profile picture delete implemented and compression used for performance
In order to delete the the profile picture, click on the profile picture
in EditProfile screen and a dialog would pop up to confirm the deletion.
Also, without a compression, I found out that the performance of the app,
especially when uploading and downloading the image file, gets horrible.
Might not be a problem in extremely faster internet connection, but
generally could be a huge problem in UX. The resolution won't compromise
neither.